Improved row action AJAX, to schedule instead of run
The fetch row action now schedules the fetch operation, rather than run it immediately.
This makes the fetching operation seem safer and more stable, decreasing the chances to timeout the AJAX for the response, allowing navigation while the feed items are being imported, and allows monitoring of the import since a cron job is is created.