Fixed minor importing bug
The algorithm that determines which feed items in the feed are "new" items (items not already stored in the database) was improperly using the feed limit.
The `$items` variable contains all the feed items, and the `$items_to_insert` contains the array slice, using the feed limit.
The `foreach` loop was disregarding the `$items_to_insert` variable, and was iterating over the original `$items` array, causing existing items to be deleted and re-imported.