The max time for the spinning icon is now 1 minute.
Each feed updates a meta field that contains the time when it started updating. This field, if found, adds the spinning icon on the Feed Sources table.
If the feed encounters an error and doesn't reach the end of the importing, this field is not removed and the icon will persist on the table.
The code in feed-processing.php deletes this field after 6 minutes.

The new code updates the meta field with each item that is imported, so that the field contains the time the last item began importing. The max time is now 30 seconds for each item.