New option `text_domain` added to `WPRSS_Help`.
This option's default value is the value of `WPRSS_Help::TEXT_DOMAIN`,
which in turn derives it's value from the core `WPRSS_TEXT_DOMAIN`
constant.

The purpost of this setting is to indicate the text domain used for any
particular tooltip. This is necessary because any other plugins may use
`WPRSS_Help` to add inline help, not just the core. However, this option
is not yet used anywhere.