WPRSS feed source bulk action selected from bottom dropdown now works.
This is because now the "action" "slug" that is processed will fall back
to the value of the "action2" field.

Unfortunately, at this time it is not possible to know which "Apply"
button was pressed, top or bottom, because none of them have the "name"
attribute set, and thus no value is sent.
This can lead to confusion, such as in the following example. If the user
selects "Activate" from the top bulk actions dropdown, then selects
"Pause" from the bottom dropdown and clicks the bottom "Apply" button, the
selected feed sources will be _activated_. Due to the described problem,
the processing algorithm has no choice but to fall back to whichever of
the 2 actions is not '-1'; the "action" field is checked before "action2".

Still, however, this fixes issue 190.