Added `WPRSS_Image_Cache` and `WPRSS_Image_Cache_Image`.
The purpose of `WPRSS_Image_Cache_Image` is to handle a particular URL and
a particular cache file, and to connect them together into a single unit.

The purpose of `WPRSS_Image_Cache` is to manage groups of cache objects as
a whole, creating, removing them, and centralizing their settings.