Activating or deactivating a license key before saving no longer causes redirect to wrong page.
The issue was due to a `_wp_http_referrer` field being included in the
resulting HTML. Because activation/deactivation happens via AJAX, the
value of this field was pointing to the AJAX script.

Removing the `_wp_http_referrer` field makes the `options.php` page
redirect back to the settings page, and thus fixes issue #215.