Changed the way feed source action nonce works (WPRACORE-411)
Instead of outputting a nonce and a referrer for every feed source, one
action nonce is output and used for the whole feed sources table. The
referrer is no longer output at all, and instead an existing referrer
field is used. The nonce field is also no longer a form field, and
therefore will not get included in search URLs. This eliminates all
previously added fields that were cluttering up the search and bulk action
URLs.

WPRACORE-411 #resolve