`#getComponentClassName()` checks root name component validity (WPRACORE-389)
If the component name specified matches a class in the root namespace,
this is only detrmined to be the resulting component name of that name
is of a valid component. Otherwise, the FQN based on the factory namespace
is returned whether a component or not.

This leads to `#createComponent()` not attempting to create instances
of invalid root namespace components.

WPRACORE-389 #resolve