Fixed quick feed actions overwriting each other
Fetching and deleting schedule WordPress cron jobs.
These cron jobs are stored in the wp_options table, as a single
entry. When multiple async requests are sent to the server, there
is a chance that some changes written to that wp_options record
will be lost. By locking the row actions until the changes are
written, the problem is slightly mitigated.